+++ App/src/components/sections/projects/Data.jsx
@@ -1,5 +1,7 @@
 import DiscordAPPLight from '../../../assets/images/app/discordapp/light.png';
 import DiscordAPPDark from '../../../assets/images/app/discordapp/dark.png';
+import PagesLight from '../../../assets/images/app/pages/light.png';
+import PagesDark from '../../../assets/images/app/pages/dark.png';
 import DauntlessBuilderLight from '../../../assets/images/contribution/dauntlessbuilder/light.png';
 import DauntlessBuilderDark from '../../../assets/images/contribution/dauntlessbuilder/dark.png';
 import NPMLight from '../../../assets/images/package/npm/light.png';
@@ -41,6 +43,24 @@ const projectsData = [
   },
   {
     id: 3,
+    category: 'app',
+    imageLight: PagesLight,
+    imageDark: PagesDark,
+    link: 'https://fjrodafo.vercel.app/',
+    title: 'Website',
+    repository: 'https://github.com/FJrodafo/Website',
+  },
+  {
+    id: 4,
+    category: 'app',
+    imageLight: PagesLight,
+    imageDark: PagesDark,
+    link: 'https://fjrodafo-cv.vercel.app/',
+    title: 'CV',
+    repository: 'https://github.com/FJrodafo/CV',
+  },
+  {
+    id: 5,
     category: 'package',
     imageLight: NPMLight,
     imageDark: NPMDark,
@@ -49,7 +69,7 @@ const projectsData = [
     repository: 'https://github.com/FJrodafo/free-port',
   },
   {
-    id: 4,
+    id: 6,
     category: 'package',
     imageLight: NPMLight,
     imageDark: NPMDark,