+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 40b7be4 - v4
+| Author: FJrodafo - Sat, 13 Jul 2024 18:09:30 +0100
+| 
 * Commit: 1cf2937 - v3
 | Author: FJrodafo - Sat, 13 Jul 2024 18:06:14 +0100
 |